backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
# gunicorn==21.2.0
# guvicorn==0.1.0

pydantic==2.6.1
pydantic-settings==2.2.1
sqlmodel==0.0.16
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
yt-dlp==2023.12.30
